+++ asyncdex/__init__.py
@@ -8,7 +8,16 @@ from .enum import (
     Relationship,
     Visibility,
 )
-from .exceptions import AsyncDexException, HTTPException, Missing, PermissionMismatch, Ratelimit, Unauthorized
+from .exceptions import (
+    AsyncDexException,
+    Captcha,
+    HTTPException,
+    InvalidCaptcha,
+    Missing,
+    PermissionMismatch,
+    Ratelimit,
+    Unauthorized,
+)
 from .list_orders import AuthorListOrder, ChapterListOrder, GroupListOrder, MangaListOrder
 from .models import *
 from .utils import AttrDict, DefaultAttrDict, InclusionExclusionPair, Interval