+++ asyncdex/models/chapter.py
@@ -347,7 +347,7 @@ class Chapter(Model, DatetimeMixin):
         await self._fetch("chapter.view", "chapter")
 
     async def load_groups(self):
-        """Shortcut method that calls :meth:`.Client.batch_groups` with the groups that belong to the group.
+        """Shortcut method that calls :meth:`.MangadexClient.batch_groups` with the groups that belong to the group.
 
         Roughly equivalent to: