+++ asyncdex/models/user.py
@@ -31,7 +31,7 @@ class User(Model):
             self.chapters = GenericModelList()
 
     async def load_chapters(self):
-        """Shortcut method that calls :meth:`.Client.batch_chapters` with the chapters that belong to the user.
+        """Shortcut method that calls :meth:`.MangadexClient.batch_chapters` with the chapters that belong to the user.
 
         Roughly equivalent to: